refactor(control-plane): split quota should-run route and packet builder
